+++ src/QuickMarc.js
@@ -3,10 +3,17 @@ import PropTypes from 'prop-types';
 import {
   Switch,
   Route,
+  useLocation,
 } from 'react-router-dom';
 
-import { IfPermission } from '@folio/stripes/core';
-import { CommandList } from '@folio/stripes/components';
+import {
+  IfPermission,
+  checkIfUserInMemberTenant,
+} from '@folio/stripes/core';
+import {
+  CommandList,
+  LoadingPane,
+} from '@folio/stripes/components';
 
 import {
   QuickMarcEditorContainer,
@@ -14,21 +21,59 @@ import {
   QuickMarcCreateWrapper,
   QuickMarcEditWrapper,
 } from './QuickMarcEditor';
+import { useUserTenantPermissions } from './queries';
 import { QUICK_MARC_ACTIONS } from './QuickMarcEditor/constants';
 import {
   MARC_TYPES,
   keyboardCommands,
 } from './common/constants';
 
+const INVALID_PERMISSION = 'invalid-permission';
+
 const QuickMarc = ({
   basePath,
   externalRecordPath,
   onClose,
+  stripes,
 }) => {
+  const location = useLocation();
+  const searchParams = new URLSearchParams(location.search);
+  const isShared = searchParams.get('shared') === 'true';
+
+  const userId = stripes?.user?.user?.id;
+  const centralTenantId = stripes.user.user?.consortium?.centralTenantId;
+  const action = location.pathname.split('/')[3];
+
+  const editMarcRecordPerm = 'ui-quick-marc.quick-marc-editor.all';
+
+  const considerCentralTenantPerm = (
+    isShared
+    && ['edit-bib'].includes(action)
+    && userId
+    && centralTenantId
+    && checkIfUserInMemberTenant(stripes)
+  );
+
+  const {
+    userPermissions: centralTenantPermissions,
+    isFetching: isCentralTenantPermissionsLoading,
+  } = useUserTenantPermissions({
+    userId,
+    tenantId: centralTenantId,
+  }, {
+    enabled: considerCentralTenantPerm,
+  });
+
+  const hasCentralTenantPerm = (perm) => {
+    return centralTenantPermissions.some(({ permissionName }) => permissionName === perm);
+  };
+
   const editorRoutesConfig = [
     {
       path: `${basePath}/edit-bib/:externalId`,
-      permission: 'ui-quick-marc.quick-marc-editor.all',
+      permission: considerCentralTenantPerm
+        ? hasCentralTenantPerm(editMarcRecordPerm) ? '' : INVALID_PERMISSION
+        : editMarcRecordPerm,
       props: {
         action: QUICK_MARC_ACTIONS.EDIT,
         wrapper: QuickMarcEditWrapper,
@@ -91,6 +136,10 @@ const QuickMarc = ({
     },
   ];
 
+  if (isCentralTenantPermissionsLoading) {
+    return <LoadingPane />;
+  }
+
   return (
     <div data-test-quick-marc>
       <CommandList
@@ -137,6 +186,7 @@ QuickMarc.propTypes = {
   basePath: PropTypes.string.isRequired,
   externalRecordPath: PropTypes.string,
   onClose: PropTypes.func.isRequired,
+  stripes: PropTypes.object.isRequired,
 };
 
 export default QuickMarc;