+++ src/queries/useLinkSuggestions/useLinkSuggestions.js
@@ -1,12 +1,18 @@
 import { useMutation } from 'react-query';
 
-import { useOkapiKy } from '@folio/stripes/core';
+import {
+  useOkapiKy,
+  useStripes,
+} from '@folio/stripes/core';
+
+import { OKAPI_TENANT_HEADER } from '../../common/constants';
 
 const useLinkSuggestions = () => {
+  const stripes = useStripes();
   const ky = useOkapiKy();
 
   const { mutateAsync, isLoading } = useMutation(
-    ({ body, isSearchByAuthorityId, ignoreAutoLinkingEnabled }) => {
+    ({ body, isSearchByAuthorityId, ignoreAutoLinkingEnabled, tenantId }) => {
       const searchParams = new URLSearchParams();
 
       if (isSearchByAuthorityId) {
@@ -20,6 +26,9 @@ const useLinkSuggestions = () => {
       return ky.post('records-editor/links/suggestion', {
         searchParams: searchParams.toString(),
         json: body,
+        headers: {
+          [OKAPI_TENANT_HEADER]: tenantId || stripes.okapi.tenant,
+        },
       }).json();
     },
   );