+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * [UIQM-501](https://issues.folio.org/browse/UIQM-501) Actualize linked fields on save.
 * [UIQM-532](https://issues.folio.org/browse/UIQM-532) *BREAKING* Bump `react` to `v18`.
 * [UIQM-480](https://issues.folio.org/browse/UIQM-480) Derive new local MARC bib record from Shared Instance record.
+* [UIQM-484](https://issues.folio.org/browse/UIQM-484) Edit MARC bib record on Shared Instance.
 
 ## [6.0.2](https://github.com/folio-org/ui-quick-marc/tree/v6.0.2) (2023-03-30)